Not Logged in trippleduck --> tripleduck, clean layout, fix localStorage undefined. install http, socket.io, socket.io-client, kafkajs, add producer, consumer, user activity logging

+++ Backend_Server/controllers/postController.js
@@ -1,17 +1,27 @@
 const Post = require("../models/Post");
 const Like = require("../models/Like");
 const Comment = require("../models/Comment");
+const { sendEvent } = require('../kafka/kafkaConfig');
 
 const postController = {};
 
 postController.createPost = async (req, res) => {
     const { authorID, authorName, content } = req.body;
     try {
-        await Post.create({
+        const newPost = await Post.create({
             authorID,
             authorName,
             content
         });
+
+        await sendEvent('user-activity', {
+            type: 'post_created',
+            userId: authorID, // Assuming you have user info in the request
+            userName: authorName,
+            postId: newPost._id,
+            timestamp: new Date().toISOString()
+        })
+
         res.sendStatus(201);
     } catch (error) {
         console.log(error);
@@ -33,18 +43,32 @@ postController.getPost = async (req, res) => {
 }
 
 postController.likePost = async (req, res) => {
-    const { user } = req.body;
+    const { userID } = req.body;
     const { postID } = req.params;
 
     try {
-        const likeCheck = await Like.findOne({ user, post: postID });
+        const likeCheck = await Like.findOne({ user: userID, post: postID });
         if (likeCheck) {
-            await Like.deleteOne({ user, post: postID });
+            await Like.deleteOne({ user: userID, post: postID });
             await Post.updateOne({ _id: postID }, { $inc: { likeCount: -1 } });
+
+            await sendEvent('user-activity', {
+                type: 'post_unliked',
+                userId: userID,
+                postId: postID,
+                timestamp: new Date().toISOString()
+            })
         }
         else {
-            await Like.create({ user, post: postID });
+            await Like.create({ user: userID, post: postID });
             await Post.updateOne({ _id: postID }, { $inc: { likeCount: 1 } });
+
+            await sendEvent('user-activity', {
+                type: 'post_liked',
+                userId: userID,
+                postId: postID,
+                timestamp: new Date().toISOString()
+            })
         }
 
         res.sendStatus(201);
@@ -59,7 +83,7 @@ postController.getPostComment = async (req, res) => {
 
     try {
         const comments = await Comment.find({ post: postID })
-            .sort({ createdAt: -1 })
+            .sort({ createdAt: -1 });
 
         res.status(200).json(comments);
     } catch (error) {
@@ -75,6 +99,16 @@ postController.commentPost = async (req, res) => {
     try {
         await Comment.create({ commenterID, commenterName, post: postID, text });
         await Post.updateOne({ _id: postID }, { $inc: { commentCount: 1 } });
+
+        await sendEvent('user-activity', {
+            type: 'post_commented',
+            userId: commenterID,
+            userName: commenterName,
+            postId: postID,
+            text: text,
+            timestamp: new Date().toISOString()
+        })
+
         res.sendStatus(201);
     } catch (error) {
         console.log(error);
